exit with failure if initial broker connection fails
The reconnect behavior of the mqqt library only kicks in after an initial connection was successful.

That meant if the broker was unavailable during agent startup, the agent got stuck in a disconnected state.

Instead of adding a loop and some backoff behavior we terminate with an error and expect the process supervisor to handle the continous retrying.